Implement video reporting - add video copy link - improve dialog UI

+++ app/src/main/java/com/examplesonly/android/adapter/ProfileVideosAdapter.java
@@ -4,35 +4,44 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import android.widget.Toast;
+
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatDialog;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.recyclerview.widget.RecyclerView.Adapter;
+
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions;
+import com.examplesonly.android.account.UserDataProvider;
 import com.examplesonly.android.component.EoAlertDialog.EoAlertDialog;
 import com.examplesonly.android.databinding.ViewExampleProfileListBinding;
 import com.examplesonly.android.handler.VideoClickListener;
 import com.examplesonly.android.model.Video;
 import com.examplesonly.android.network.Api;
+import com.examplesonly.android.network.user.UserInterface;
 import com.examplesonly.android.network.video.VideoInterface;
 import com.examplesonly.android.util.MediaUtil;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
+
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 import timber.log.Timber;
+
 public class ProfileVideosAdapter extends Adapter<ProfileVideosAdapter.VIewHolder> {
 
     private final ArrayList<Video> videos;
     private final Context context;
     private final LayoutInflater inflater;
     private final VideoClickListener clickListener;
+    private boolean isLoggedInUser = false;
     private VideoInterface mVideoInterface;
+    private UserDataProvider userDataProvider;
 
     public ProfileVideosAdapter(ArrayList<Video> videos, Context context, VideoClickListener clickListener) {
         this.videos = videos;
@@ -40,8 +49,13 @@ public class ProfileVideosAdapter extends Adapter<ProfileVideosAdapter.VIewHolde
         this.clickListener = clickListener;
 
         mVideoInterface = new Api(context).getClient().create(VideoInterface.class);
+        userDataProvider = UserDataProvider.getInstance(context);
 
         inflater = LayoutInflater.from(context);
+    }
+
+    public void setIsLoggedInUser(Boolean isLoggedInUser) {
+        this.isLoggedInUser = isLoggedInUser;
     }
 
     @NonNull
@@ -91,42 +105,44 @@ public class ProfileVideosAdapter extends Adapter<ProfileVideosAdapter.VIewHolde
                 clickListener.onVideoClicked(video);
             });
 
-            mChooseCategoryBinding.imageCard.setOnLongClickListener(v -> {
-                EoAlertDialog deleteDialog = new EoAlertDialog(context)
-                        .setTitle("Delete")
-                        .setPositiveText("Delete")
-                        .setNegativeText("Cancel")
-                        .setPositiveClickListener(dialog -> {
-                            dialog.dismiss();
-                            mVideoInterface.deleteVideo(video.getVideoId()).enqueue(
-                                    new Callback<HashMap<String, String>>() {
-                                        @Override
-                                        public void onResponse(final Call<HashMap<String, String>> call,
-                                                final Response<HashMap<String, String>> response) {
-                                            if (response.isSuccessful()) {
-                                                Toast.makeText(context, "Video deleted", Toast.LENGTH_SHORT).show();
-                                            } else {
-                                                try {
-                                                    Timber.e("deleteVideo error %s", response.errorBody().string());
-                                                } catch (IOException e) {
-                                                    e.printStackTrace();
+            if (isLoggedInUser)
+                mChooseCategoryBinding.imageCard.setOnLongClickListener(v -> {
+                    EoAlertDialog deleteDialog = new EoAlertDialog(context)
+                            .setTitle("Delete")
+                            .setPositiveText("Delete")
+                            .setNegativeText("Cancel")
+                            .setPositiveClickListener(dialog -> {
+                                dialog.dismiss();
+                                mVideoInterface.deleteVideo(video.getVideoId()).enqueue(
+                                        new Callback<HashMap<String, String>>() {
+                                            @Override
+                                            public void onResponse(final Call<HashMap<String, String>> call,
+                                                                   final Response<HashMap<String, String>> response) {
+                                                if (response.isSuccessful()) {
+                                                    Toast.makeText(context, "Video deleted", Toast.LENGTH_SHORT).show();
+                                                } else {
+                                                    try {
+                                                        Timber.e("deleteVideo error %s", response.errorBody().string());
+                                                    } catch (IOException e) {
+                                                        e.printStackTrace();
+                                                    }
+                                                    Toast.makeText(context, "Could not delete video", Toast.LENGTH_SHORT).show();
                                                 }
+                                            }
+
+                                            @Override
+                                            public void onFailure(final Call<HashMap<String, String>> call,
+                                                                  final Throwable t) {
                                                 Toast.makeText(context, "Could not delete video", Toast.LENGTH_SHORT).show();
                                             }
-                                        }
+                                        });
+                            })
+                            .setNegativeClickListener(AppCompatDialog::dismiss);
 
-                                        @Override
-                                        public void onFailure(final Call<HashMap<String, String>> call,
-                                                final Throwable t) {
-                                            Toast.makeText(context, "Could not delete video", Toast.LENGTH_SHORT).show();
-                                        }
-                                    });
-                        })
-                        .setNegativeClickListener(AppCompatDialog::dismiss);
+                    deleteDialog.show();
+                    return false;
+                });
 
-                deleteDialog.show();
-                return false;
-            });
         }
     }
 }
